+++ src/apps/__init__.py
@@ -2,7 +2,7 @@ from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 from src.apps.user.controller import user
 from src.apps.account.controller import account
-from src.libs.error_handler import error_handlers
+from src.libs.api.error_handler import error_handlers
 
 
 def create_app():
