+++ src/apps/user/repository.py
@@ -1,5 +1,5 @@
-from src.libs.exception import DBError
-from src.libs.error_code import DBErrorCode
+from src.libs.api.exception import DBError
+from src.libs.api.error_code import DBErrorCode
 
 
 class UserRepository:
