+++ src/libs/api/exception.py
