+++ src/libs/api/util.py
@@ -2,6 +2,7 @@ import os
 from datetime import datetime
 from fastapi import UploadFile
 
+
 def create_folder_if_not_exists(folder_path: str) -> str:
     if not os.path.exists(folder_path):
         os.makedirs(folder_path)
