+++ tests/libs/test_error_code.py
@@ -1,4 +1,4 @@
-from src.libs.error_code import DBErrorCode
+from src.libs.api.error_code import DBErrorCode
 
 
 def test_DBErrorCode_enum_class():
